requirements.txt
aiofiles==0.8.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2
cryptography==36.0.1
decorator==5.1.1
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.61.1
h11==0.9.0
httplib2==0.20.2
httptools==0.1.2
idna==3.3
jwt==1.3.1
motor==2.2.0
oauth2==1.9.0.post1
passlib==1.7.4
Pillow==9.0.1
pycparser==2.21routes
pydantic==1.9.0
PyJWT==2.3.0
pymongo==3.12.3
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
requests==2.27.1
six==1.16.0
starlette==0.13.6
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.11.8
uvloop==0.16.0
validators==0.18.2
websockets==8.1